# Initial Span support without performance regression