--- a/.gitignore
+++ b/.gitignore
@@ -250,3 +250,4 @@ paket-files/
 # JetBrains Rider
 .idea/
 *.sln.iml
+/MetroHash.Benchmark/BenchmarkDotNet.Artifacts/results