--- a/MetroHash.Benchmark/MetroBenchmark.cs
+++ b/MetroHash.Benchmark/MetroBenchmark.cs
@@ -1,4 +1,5 @@
-﻿using System.Text;
+﻿using System;
+using System.Text;
 using BenchmarkDotNet.Attributes;
 
 namespace MetroHash.Benchmark
@@ -9,6 +10,7 @@ namespace MetroHash.Benchmark
         private static readonly byte[] Data =
             Encoding.ASCII.GetBytes("012345678901234567890123456789012345678901234567890123456789012");
 
+        private static readonly byte[] Buffer = new byte[16];
         [Benchmark]
         public byte[] MetroHash128NonIncremental()
         {
@@ -21,6 +23,23 @@ namespace MetroHash.Benchmark
             var metroHash = new MetroHash128(0);
             metroHash.Update(Data, 0, Data.Length);
             return metroHash.FinalizeHash();
+        }
+
+
+        [Benchmark]
+        public byte[] MetroHash128NonIncrementalSpan()
+        {
+            MetroHash128.Hash(0, Data, Buffer);
+            return Buffer;
+        }
+
+        [Benchmark]
+        public byte[] MetroHash128IncrementalSpan()
+        {
+            var metroHash = new MetroHash128(0);
+            metroHash.Update(Data);
+            metroHash.FinalizeHash(Buffer);
+            return Buffer;
         }
     }
 }