--- a/MetroHash.Benchmark/MetroHash.Benchmark.csproj
+++ b/MetroHash.Benchmark/MetroHash.Benchmark.csproj
@@ -2,6 +2,7 @@
   <PropertyGroup>
     <TargetFrameworks>netcoreapp2.1;net461</TargetFrameworks>
     <OutputType>Exe</OutputType>
+    <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <ItemGroup>
     <Compile Remove="BenchmarkDotNet.Artifacts\**" />