--- a/MetroHash.Tests/Tests.cs
+++ b/MetroHash.Tests/Tests.cs
@@ -32,6 +32,7 @@ namespace MetroHash.Tests
                 remaining -= toCopy;
                 offset += toCopy;
             }
+
             var hash = metroHash.FinalizeHash();
             var comparison = new byte[]
             {
@@ -85,6 +86,23 @@ namespace MetroHash.Tests
             Assert.Equal(comparison, hash);
         }
 
+
+        [Fact]
+        public void IncrementalSeed0Span()
+        {
+            var input = Encoding.ASCII.GetBytes(TestString);
+            var metroHash = new MetroHash128(0);
+            metroHash.Update(input);
+            Span<byte> span = stackalloc byte[16];
+            metroHash.FinalizeHash(span);
+            var comparison = new byte[]
+            {
+                0xC7, 0x7C, 0xE2, 0xBF, 0xA4, 0xED, 0x9F, 0x9B,
+                0x05, 0x48, 0xB2, 0xAC, 0x50, 0x74, 0xA2, 0x97
+            };
+            Assert.True(span.SequenceEqual(comparison));
+        }
+
         [Fact]
         public void IncrementalSeed1()
         {
@@ -117,6 +135,25 @@ namespace MetroHash.Tests
             Assert.Equal(comparison, hash);
         }
 
+        [Fact]
+        public void IncrementalSeed1SingleByteSpan()
+        {
+            var input = Encoding.ASCII.GetBytes(TestString);
+            var metroHash = new MetroHash128(1);
+            for (var i = 0; i < input.Length; i++)
+            {
+                metroHash.Update(input.AsSpan().Slice(i, 1));
+            }
+
+            var hash = metroHash.FinalizeHash();
+            var comparison = new byte[]
+            {
+                0x45, 0xA3, 0xCD, 0xB8, 0x38, 0x19, 0x9D, 0x7F,
+                0xBD, 0xD6, 0x8D, 0x86, 0x7A, 0x14, 0xEC, 0xEF
+            };
+            Assert.Equal(comparison, hash);
+        }
+
 
         [Fact]
         public void IncrementalSeed1SingleByteOffset()
@@ -135,16 +172,32 @@ namespace MetroHash.Tests
         }
 
         [Fact]
+        public void NonIncrementalSeed0Span()
+        {
+            var input = Encoding.ASCII.GetBytes(TestString);
+            Span<byte> span = stackalloc byte[16];
+            MetroHash128.Hash(0, input, span);
+            var comparison = new byte[]
+            {
+                0xC7, 0x7C, 0xE2, 0xBF, 0xA4, 0xED, 0x9F, 0x9B,
+                0x05, 0x48, 0xB2, 0xAC, 0x50, 0x74, 0xA2, 0x97
+            };
+            Assert.True(span.SequenceEqual(comparison));
+        }
+
+        [Fact]
         public void NonIncrementalSeed0()
         {
             var input = Encoding.ASCII.GetBytes(TestString);
             var hash = MetroHash128.Hash(0, input, 0, input.Length);
+            var secondHash = OldMetroHash128.Hash(0, input, 0, input.Length);
             var comparison = new byte[]
             {
                 0xC7, 0x7C, 0xE2, 0xBF, 0xA4, 0xED, 0x9F, 0x9B,
                 0x05, 0x48, 0xB2, 0xAC, 0x50, 0x74, 0xA2, 0x97
             };
             Assert.Equal(comparison, hash);
+            Assert.Equal(comparison, secondHash);
         }
 
         [Fact]
@@ -179,6 +232,27 @@ namespace MetroHash.Tests
                 Assert.Equal(firstHash, secondHash);
                 var thirdHash = OldMetroHash128.Hash(0, input, 0, input.Length);
                 Assert.Equal(firstHash, thirdHash);
+            }
+        }
+
+        [Fact]
+        public void RandomSizesSpan()
+        {
+            for (var i = 0; i < 1000; i++)
+            {
+                var prng = new Random();
+                var size = prng.Next(10000);
+                var input = new byte[size];
+                prng.NextBytes(input);
+                Span<byte> firstOutput = stackalloc byte[16];
+                MetroHash128.Hash(0, input, firstOutput);
+                var metroHash = new MetroHash128(0);
+                metroHash.Update(input);
+                Span<byte> secondOutput = stackalloc byte[16];
+                metroHash.FinalizeHash(secondOutput);
+                Assert.True(firstOutput.SequenceEqual(secondOutput));
+                var thirdHash = OldMetroHash128.Hash(0, input, 0, input.Length);
+                Assert.True(firstOutput.SequenceEqual(thirdHash));
             }
         }
     }