… MetroHash.Benchmark/App.config
… MetroHash.Benchmark/Properties/AssemblyInfo.cs
… MetroHash.Benchmark/packages.config
… MetroHash.Tests/Properties/AssemblyInfo.cs
… MetroHash.Tests/packages.config