The MetroHash128IncrementalSpan  copies the hash into an output span, the DirectReturn uses the internal buffer directly as a return value.
``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-4790K CPU 4.00GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
Frequency=3906246 Hz, Resolution=256.0003 ns, Timer=TSC
.NET Core SDK=2.1.300-rtm-008701
  [Host]     : .NET Core 2.1.0-rtm-26502-02 (CoreCLR 4.6.26502.03, CoreFX 4.6.26502.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-rtm-26502-02 (CoreCLR 4.6.26502.03, CoreFX 4.6.26502.02), 64bit RyuJIT


```
|                                  Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
|---------------------------------------- |---------:|----------:|----------:|-------:|----------:|
|              MetroHash128NonIncremental | 43.50 ns | 0.8945 ns | 0.9571 ns | 0.0095 |      40 B |
|                 MetroHash128Incremental | 65.80 ns | 0.5214 ns | 0.4354 ns | 0.0380 |     160 B |
|          MetroHash128NonIncrementalSpan | 45.72 ns | 0.5362 ns | 0.5016 ns |      - |       0 B |
|             MetroHash128IncrementalSpan | 78.18 ns | 1.6029 ns | 1.8459 ns | 0.0380 |     160 B |
| MetroHash128IncrementalSpanDirectReturn | 67.61 ns | 0.2297 ns | 0.1918 ns | 0.0380 |     160 B |